Temporary adding copy of nerc-mo agreement 'gov'
This is a copy of the NERC-Met Office agreement used for most Met Office dataset access. This is presently also used for 'government' use access route that is supported by CEDA. Adding in this _gov use copy so that it can be marked up as 'government' use in the accessCtl system to allow roll out of the multiple-licence route whist a different licence arrangement is agreed with the Met Office.